dashboard/pyproject.toml
[project]
name = "nyc-citibike-dashboard"
version = "0.1.0"
description = "NYC Citi Bike Trips Dashboard"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dash[cloud]>=4.0.0",
    "google-cloud-bigquery>=3.40.1",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "plotly>=6.6.0",
    "polars>=1.39.0",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.2",
]
